Updated Docs and Recycle FDTD session
-The docs have been updated for the broadband figure of merit. Please
note that there is an API change and old code using lumopt will need to
be updated to run. This involves adding some new arguments when defining
a ModeMatch object to define the broaband target function.
-The FDTD Solutions session is recycled so only one session per
optimization is created. This improved performance by avoiding the load
time for the application

This commit corresponds to the lumopt version packaged inside FDTD
Solutions 2019a R4 (8.21.1882)