Added topology optimization and enhancements
Added a major new feature to perform topology optimization. This
supports 2D optimization and 3D optimization with an extruded quasi-2D
layer. The topology optimization is constrained to binary materials
and respects a minimum feature size which the user can control.

Added a number of performance enhancements to the shape based
optimization to support grating coupler devices with large numbers of
parameters and using the high presision volume averrage mesh refnement
strategy.

Example files have been added and updated to demonstrate the new
features included in this commit.

This commit corresponds to the version shipped with Lumerical FDTD 2019b
R2